+++ common/src/main/java/com/nb6868/onex/common/log/BaseLogService.java
@@ -6,9 +6,11 @@ import org.springframework.stereotype.Service;
 public interface BaseLogService {
 
     /**
-     * 保存到数据库
+     * 保存日志
+     * 做分发，按日志类型和保存方式分别做保存
+     *
      * @param log 日志内容
      */
-    void saveToDb(LogBody log);
+    void saveLog(LogBody log);
 
 }